Move components into src, tests into client
